Introduce inheritance in data structure elements
