+++ osm2gtfs/creators/accra/trips_creator_accra.py
@@ -4,7 +4,7 @@ from datetime import timedelta, datetime
 
 from osm2gtfs.creators.trips_creator import TripsCreator
 from osm2gtfs.core.helper import Helper
-from osm2gtfs.core.routes import Line
+from osm2gtfs.core.elements import Line
 
 
 class TripsCreatorAccra(TripsCreator):
